libbalde/httpd: downgraded to http/1.0
it is too hard to guess the end of an http/1.1 request, even with the
help of http-parser. replaced old buggy code with a simple homemade
http/1.0 server/parser that does the trick quite well for a simple
development webserver :)

and for the sake of simplicity, this server will ignore the http request
body if the client does not provides a valid content-length header.