scgi: added scgi protocol support.
this patch implements scgi protocol, as described in the specification:
http://python.ca/scgi/protocol.txt

scgi protocol is way simpler than fastcgi, and easier to implement
without external libraries like fcgi, that is unmantained.

some webservers like nginx does not support PATH_INFO properly and a
workaround was implemented in this patch, but it fails if the
application is deployed to a subdirectory, instead of domain root.

current suggestion is to use lighttpd, or deploy apps to domain root,
until nginx' scgi module accepts an option similar to
fastcgi_split_path_info from its fastcgi module.