sapi: implemented fastcgi support
this patch adds fastcgi support, without using any external library
other than glib, that is already a dependency.

this patch lacks tests and would deserve a blog post, but I won't
have time and/or energy for it any time soon, then I'm just commiting
what I did last weekend, as is, and adding a few comments here.

i'm not using the fastcgi sdk because it is unmaintained. even the
fastcgi spec is not available from the "official" source at this point.

some design details:

- only the responder role is implemented.
- supports multiplexed requests.
- there's (a lot of) room for memory usage optimizations.

and last but not least, i wrote this because someone told me on hackernews
that scgi is not a good option for production deployments, because it
opens a new connection between the webserver and the application for each
client request. while this is true, the same happens for fastcgi in the
default setup for some of the major web servers, like nginx, and as far
as I researched, the usage of upstream keepalive and fastcgi_keep_conn
is not really wide-spread. it is good to have it supported anyway.